+++ typespec-tests/src/main/java/com/cadl/armresourceprovider/models/CustomTemplateResource.java
@@ -5,6 +5,7 @@
 package com.cadl.armresourceprovider.models;
 
 import com.azure.core.management.Region;
+import com.azure.core.management.SystemData;
 import com.azure.core.util.Context;
 import com.cadl.armresourceprovider.fluent.models.CustomTemplateResourceInner;
 import java.util.Map;
@@ -49,18 +50,32 @@ public interface CustomTemplateResource {
     Map<String, String> tags();
 
     /**
-     * Gets the properties property: The resource-specific properties for this resource.
-     * 
-     * @return the properties value.
-     */
-    CustomTemplateResourceProperties properties();
-
-    /**
      * Gets the identity property: Managed identity.
      * 
      * @return the identity value.
      */
     ManagedIdentityProperties identity();
+
+    /**
+     * Gets the systemData property: Azure Resource Manager metadata containing createdBy and modifiedBy information.
+     * 
+     * @return the systemData value.
+     */
+    SystemData systemData();
+
+    /**
+     * Gets the provisioningState property: The status of the last operation.
+     * 
+     * @return the provisioningState value.
+     */
+    ProvisioningState provisioningState();
+
+    /**
+     * Gets the propertyRemovedInStable property: The propertyRemovedInStable property.
+     * 
+     * @return the propertyRemovedInStable value.
+     */
+    String propertyRemovedInStable();
 
     /**
      * Gets the region of the resource.
@@ -145,8 +160,9 @@ public interface CustomTemplateResource {
          * The stage of the CustomTemplateResource definition which contains all the minimum required properties for the
          * resource to be created, but also allows for any other optional properties to be specified.
          */
-        interface WithCreate extends DefinitionStages.WithTags, DefinitionStages.WithProperties,
-            DefinitionStages.WithIdentity, DefinitionStages.WithIfMatch, DefinitionStages.WithIfNoneMatch {
+        interface WithCreate extends DefinitionStages.WithTags, DefinitionStages.WithIdentity,
+            DefinitionStages.WithPropertyRemovedInStable, DefinitionStages.WithIfMatch,
+            DefinitionStages.WithIfNoneMatch {
             /**
              * Executes the create request.
              * 
@@ -177,19 +193,6 @@ public interface CustomTemplateResource {
         }
 
         /**
-         * The stage of the CustomTemplateResource definition allowing to specify properties.
-         */
-        interface WithProperties {
-            /**
-             * Specifies the properties property: The resource-specific properties for this resource..
-             * 
-             * @param properties The resource-specific properties for this resource.
-             * @return the next definition stage.
-             */
-            WithCreate withProperties(CustomTemplateResourceProperties properties);
-        }
-
-        /**
          * The stage of the CustomTemplateResource definition allowing to specify identity.
          */
         interface WithIdentity {
@@ -200,6 +203,19 @@ public interface CustomTemplateResource {
              * @return the next definition stage.
              */
             WithCreate withIdentity(ManagedIdentityProperties identity);
+        }
+
+        /**
+         * The stage of the CustomTemplateResource definition allowing to specify propertyRemovedInStable.
+         */
+        interface WithPropertyRemovedInStable {
+            /**
+             * Specifies the propertyRemovedInStable property: The propertyRemovedInStable property..
+             * 
+             * @param propertyRemovedInStable The propertyRemovedInStable property.
+             * @return the next definition stage.
+             */
+            WithCreate withPropertyRemovedInStable(String propertyRemovedInStable);
         }
 
         /**